+++ src/spatialdata/_core/operations/map.py
@@ -1,20 +1,15 @@
 from __future__ import annotations
 
-import math
-import operator
-from collections.abc import Iterable, Mapping
-from functools import reduce
+from collections.abc import Callable, Iterable, Mapping
 from types import MappingProxyType
-from typing import TYPE_CHECKING, Any, Callable
+from typing import TYPE_CHECKING, Any
 
 import dask.array as da
 import numpy as np
 from dask.array.overlap import coerce_depth
-from datatree import DataTree
-from xarray import DataArray
+from xarray import DataArray, DataTree
 
-from spatialdata._types import IntArray
-from spatialdata.models._utils import get_axes_names, get_channels, get_raster_model_from_data_dims
+from spatialdata.models._utils import get_axes_names, get_channel_names, get_raster_model_from_data_dims
 from spatialdata.transformations import get_transformation
 
 __all__ = ["map_raster"]
@@ -133,7 +128,7 @@ def map_raster(
 
     if "c" in dims:
         if c_coords is None:
-            c_coords = range(arr.shape[0]) if arr.shape[0] != len(get_channels(data)) else get_channels(data)
+            c_coords = range(arr.shape[0]) if arr.shape[0] != len(get_channel_names(data)) else get_channel_names(data)
     else:
         c_coords = None
     if transformations is None: