+++ src/spatialdata/_io/_utils.py
@@ -16,12 +16,11 @@ import zarr
 from anndata import AnnData
 from dask.array import Array as DaskArray
 from dask.dataframe import DataFrame as DaskDataFrame
-from datatree import DataTree
 from geopandas import GeoDataFrame
-from xarray import DataArray
+from xarray import DataArray, DataTree
 
 from spatialdata._core.spatialdata import SpatialData
-from spatialdata._utils import iterate_pyramid_levels
+from spatialdata._utils import get_pyramid_levels
 from spatialdata.models._utils import (
     MappingToCoordinateSystem_t,
     SpatialElement,
@@ -53,7 +52,7 @@ def _get_transformations_from_ngff_dict(
     list_of_ngff_transformations = [NgffBaseTransformation.from_dict(d) for d in list_of_encoded_ngff_transformations]
     list_of_transformations = [BaseTransformation.from_ngff(t) for t in list_of_ngff_transformations]
     transformations = {}
-    for ngff_t, t in zip(list_of_ngff_transformations, list_of_transformations):
+    for ngff_t, t in zip(list_of_ngff_transformations, list_of_transformations, strict=True):
         assert ngff_t.output_coordinate_system is not None
         transformations[ngff_t.output_coordinate_system.name] = t
     return transformations
@@ -105,6 +104,27 @@ def overwrite_coordinate_transformations_raster(
     group.attrs["multiscales"] = multiscales
 
 
+def overwrite_channel_names(group: zarr.Group, element: DataArray | DataTree) -> None:
+    """Write channel metadata to a group."""
+    if isinstance(element, DataArray):
+        channel_names = element.coords["c"].data.tolist()
+    else:
+        channel_names = element["scale0"]["image"].coords["c"].data.tolist()
+
+    channel_metadata = [{"label": name} for name in channel_names]
+    omero_meta = group.attrs["omero"]
+    omero_meta["channels"] = channel_metadata
+    group.attrs["omero"] = omero_meta
+    multiscales_meta = group.attrs["multiscales"]
+    if len(multiscales_meta) != 1:
+        raise ValueError(
+            f"Multiscale metadata must be of length one but got length {len(multiscales_meta)}. Data might"
+            f"be corrupted."
+        )
+    multiscales_meta[0]["metadata"]["omero"]["channels"] = channel_metadata
+    group.attrs["multiscales"] = multiscales_meta
+
+
 def _write_metadata(
     group: zarr.Group,
     group_type: str,
@@ -120,22 +140,6 @@ def _write_metadata(
     # them with overwrite_coordinate_transformations_non_raster()
     group.attrs["coordinateTransformations"] = []
     group.attrs["spatialdata_attrs"] = attrs
-
-
-def _iter_multiscale(
-    data: DataTree,
-    attr: str | None,
-) -> list[Any]:
-    # TODO: put this check also in the validator for raster multiscales
-    for i in data:
-        variables = set(data[i].variables.keys())
-        names: set[str] = variables.difference({"c", "z", "y", "x"})
-        if len(names) != 1:
-            raise ValueError(f"Invalid variable name: `{names}`.")
-    name: str = next(iter(names))
-    if attr is not None:
-        return [getattr(data[i][name], attr) for i in data]
-    return [data[i][name] for i in data]
 
 
 class dircmp(filecmp.dircmp):  # type: ignore[type-arg]
@@ -229,7 +233,7 @@ def get_dask_backing_files(element: SpatialData | SpatialElement | AnnData) -> l
 def _(element: SpatialData) -> list[str]:
     files: set[str] = set()
     for e in element._gen_spatial_element_values():
-        if isinstance(e, (DataArray, DataTree, DaskDataFrame)):
+        if isinstance(e, DataArray | DataTree | DaskDataFrame):
             files = files.union(get_dask_backing_files(e))
     return list(files)
 
@@ -241,8 +245,8 @@ def _(element: DataArray) -> list[str]:
 
 @get_dask_backing_files.register(DataTree)
 def _(element: DataTree) -> list[str]:
-    xdata0 = next(iter(iterate_pyramid_levels(element)))
-    return _get_backing_files(xdata0.data)
+    dask_data_scale0 = get_pyramid_levels(element, attr="data", n=0)
+    return _get_backing_files(dask_data_scale0)
 
 
 @get_dask_backing_files.register(DaskDataFrame)