+++ src/spatialdata/datasets.py
@@ -10,13 +10,12 @@ import pandas as pd
 import scipy
 from anndata import AnnData
 from dask.dataframe import DataFrame as DaskDataFrame
-from datatree import DataTree
 from geopandas import GeoDataFrame
 from numpy.random import default_rng
 from shapely.affinity import translate
 from shapely.geometry import MultiPolygon, Point, Polygon
 from skimage.segmentation import slic
-from xarray import DataArray
+from xarray import DataArray, DataTree
 
 from spatialdata._core.operations.aggregate import aggregate
 from spatialdata._core.query.relational_query import get_element_instances
@@ -41,7 +40,7 @@ def blobs(
     n_shapes: int = 5,
     extra_coord_system: str | None = None,
     n_channels: int = 3,
-    c_coords: ArrayLike | None = None,
+    c_coords: str | list[str] | None = None,
 ) -> SpatialData:
     """
     Blobs dataset.
@@ -109,7 +108,7 @@ class BlobsDataset:
         n_shapes: int = 5,
         extra_coord_system: str | None = None,
         n_channels: int = 3,
-        c_coords: ArrayLike | None = None,
+        c_coords: str | list[str] | None = None,
     ) -> None:
         """
         Blobs dataset.
@@ -177,13 +176,13 @@ class BlobsDataset:
         transformations: dict[str, Any] | None = None,
         length: int = 512,
         n_channels: int = 3,
-        c_coords: ArrayLike | None = None,
+        c_coords: str | list[str] | None = None,
         multiscale: bool = False,
     ) -> DataArray | DataTree:
         masks = []
         for i in range(n_channels):
             mask = self._generate_blobs(length=length, seed=i)
-            mask = (mask - mask.min()) / mask.ptp()  # type: ignore[attr-defined]
+            mask = (mask - mask.min()) / np.ptp(mask)  # type: ignore[attr-defined]
             masks.append(mask)
 
         x = np.stack(masks, axis=0)