+++ tests/core/operations/test_rasterize.py
@@ -6,17 +6,16 @@ import pandas as pd
 import pytest
 from anndata import AnnData
 from dask.dataframe import DataFrame as DaskDataFrame
-from datatree import DataTree
 from geopandas import GeoDataFrame
 from multiscale_spatial_image import MultiscaleSpatialImage
 from shapely import MultiPolygon, box
 from spatial_image import SpatialImage
-from xarray import DataArray
+from xarray import DataArray, DataTree
 
 from spatialdata import SpatialData, get_extent
 from spatialdata._core.operations.rasterize import rasterize
 from spatialdata._core.query.relational_query import get_element_instances
-from spatialdata._io._utils import _iter_multiscale
+from spatialdata._utils import get_pyramid_levels
 from spatialdata.models import PointsModel, ShapesModel, TableModel, get_axes_names
 from spatialdata.models._utils import get_spatial_axes
 from spatialdata.transformations import MapAxis
@@ -57,7 +56,7 @@ def test_rasterize_raster(_get_raster):
         if isinstance(raster, DataArray):
             return raster.data.compute()
 
-        xdata = next(iter(_iter_multiscale(raster, None)))
+        xdata = get_pyramid_levels(raster, n=0)
         return xdata.data.compute()
 
     for element_name, raster in rasters.items():